Migrate full-stack UVCC test
