+++ test/circuit/library/ansatzes/test_uvcc.py
@@ -14,13 +14,21 @@
 
 from test import QiskitNatureTestCase
 
+import unittest
+
 from ddt import ddt, data, unpack
 
+from qiskit import BasicAer
+from qiskit.utils import QuantumInstance, algorithm_globals
+from qiskit.algorithms import VQE
+from qiskit.algorithms.optimizers import COBYLA
 from qiskit_nature.circuit.library.ansatzes import UVCC
+from qiskit_nature.circuit.library.initial_states import VSCF
 from qiskit_nature.mappers.second_quantization import DirectMapper
 from qiskit_nature.operators.second_quantization import VibrationalOp
 from qiskit_nature.operators.second_quantization.qubit_converter import QubitConverter
-
+from qiskit_nature.problems.second_quantization.vibrational.vibrational_label_builder import \
+    _create_labels
 
 def assert_ucc_like_ansatz(test_case, ansatz, num_modals, expected_ops):
     """Assertion utility."""
@@ -41,9 +49,9 @@ class TestUVCC(QiskitNatureTestCase):
 
     @unpack
     @data(
-        ('s', [2], [VibrationalOp([('+-', 1j), ('-+', 1j)], 1, 2)]),
-        ('s', [2, 2], [VibrationalOp([('+-II', 1j), ('-+II', 1j)], 2, 2),
-                       VibrationalOp([('II+-', 1j), ('II-+', 1j)], 2, 2)]),
+        ('s', [2], [VibrationalOp([('+-', 1j), ('-+', -1j)], 1, 2)]),
+        ('s', [2, 2], [VibrationalOp([('+-II', 1j), ('-+II', -1j)], 2, 2),
+                       VibrationalOp([('II+-', 1j), ('II-+', -1j)], 2, 2)]),
     )
     def test_ucc_ansatz(self, excitations, num_modals, expect):
         """Tests the UVCC Ansatz."""
@@ -54,3 +62,65 @@ class TestUVCC(QiskitNatureTestCase):
                       excitations=excitations)
 
         assert_ucc_like_ansatz(self, ansatz, num_modals, expect)
+
+
+class TestUVCCVSCF(QiskitNatureTestCase):
+    """Test for these extensions."""
+
+    def setUp(self):
+        super().setUp()
+        algorithm_globals.random_seed = 8
+        self.reference_energy = 592.5346633819712
+
+    def test_uvcc_vscf(self):
+        """ uvcc vscf test """
+
+        co2_2modes_2modals_2body = [[[[[0, 0, 0]], 320.8467332810141],
+                                     [[[0, 1, 1]], 1760.878530705873],
+                                     [[[1, 0, 0]], 342.8218290247543],
+                                     [[[1, 1, 1]], 1032.396323618631]],
+                                    [[[[0, 0, 0], [1, 0, 0]], -57.34003649795117],
+                                     [[[0, 0, 1], [1, 0, 0]], -56.33205925807966],
+                                     [[[0, 1, 0], [1, 0, 0]], -56.33205925807966],
+                                     [[[0, 1, 1], [1, 0, 0]], -60.13032761856809],
+                                     [[[0, 0, 0], [1, 0, 1]], -65.09576309934431],
+                                     [[[0, 0, 1], [1, 0, 1]], -62.2363839133389],
+                                     [[[0, 1, 0], [1, 0, 1]], -62.2363839133389],
+                                     [[[0, 1, 1], [1, 0, 1]], -121.5533969109279],
+                                     [[[0, 0, 0], [1, 1, 0]], -65.09576309934431],
+                                     [[[0, 0, 1], [1, 1, 0]], -62.2363839133389],
+                                     [[[0, 1, 0], [1, 1, 0]], -62.2363839133389],
+                                     [[[0, 1, 1], [1, 1, 0]], -121.5533969109279],
+                                     [[[0, 0, 0], [1, 1, 1]], -170.744837386338],
+                                     [[[0, 0, 1], [1, 1, 1]], -167.7433236025723],
+                                     [[[0, 1, 0], [1, 1, 1]], -167.7433236025723],
+                                     [[[0, 1, 1], [1, 1, 1]], -179.0536532281924]]]
+        num_modes = 2
+        num_modals = [2, 2]
+
+        vibrational_op_labels = _create_labels(co2_2modes_2modals_2body)
+        vibr_op = VibrationalOp(vibrational_op_labels, num_modes, num_modals)
+
+        converter = QubitConverter(DirectMapper())
+
+        qubit_op = converter.to_qubit_ops([vibr_op])[0]
+
+        init_state = VSCF(num_modals)
+
+        uvcc_varform = UVCC(converter, num_modals, 'sd', initial_state=init_state)
+        print(uvcc_varform.draw())
+
+        q_instance = QuantumInstance(BasicAer.get_backend('statevector_simulator'),
+                                     seed_transpiler=90, seed_simulator=12)
+        optimizer = COBYLA(maxiter=1000)
+
+        algo = VQE(uvcc_varform, optimizer=optimizer, quantum_instance=q_instance)
+        vqe_result = algo.compute_minimum_eigenvalue(qubit_op)
+
+        energy = vqe_result.optimal_value
+
+        self.assertAlmostEqual(energy, self.reference_energy, places=4)
+
+
+if __name__ == '__main__':
+    unittest.main()
